ai_gongwu_backend/requirements.txt
fastapi
openai
openai-whisper
opencv-python-headless
pydantic
pydantic-settings
python-multipart
SQLAlchemy
torch
uvicorn[standard]
